+++ src/main/java/umk/zychu/inzynierka/repository/OrlikDAO.java
@@ -3,13 +3,12 @@ package umk.zychu.inzynierka.repository;
 import java.util.List;
 import java.util.Map;
 
-import umk.zychu.inzynierka.model.GraphicEntity;
 import umk.zychu.inzynierka.model.Orlik;
 
 
 
 public interface OrlikDAO {
-	Orlik getOrlikById(int id);
+	Orlik getOrlikById(long id);
 	List<Orlik> getOrliks();
 	void saveOrlik(Orlik orlik);
 	Map<String, String> getOrliksIdsAndNames();
